# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "FilesystemEntriesResponse": {
        "properties": {
          "can_go_up": {
            "title": "Can Go Up",
            "type": "boolean"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FilesystemEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "root": {
            "title": "Root",
            "type": "string"
          }
        },
        "required": [
          "root",
          "path",
          "parent",
          "can_go_up",
          "entries"
        ],
        "title": "FilesystemEntriesResponse",
        "type": "object"
      },
      "FilesystemEntryResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "selectable": {
            "title": "Selectable",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "directory",
              "file"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "type",
          "selectable"
        ],
        "title": "FilesystemEntryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LLMConfig": {
        "description": "Root LLM configuration model.\n\nValidates the complete llm-config.yml structure with schema versioning.",
        "properties": {
          "default_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default profile name when none specified",
            "title": "Default Profile"
          },
          "profiles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProfileConfig"
            },
            "description": "Profile tier definitions (reference providers)",
            "title": "Profiles",
            "type": "object"
          },
          "providers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProviderConfig"
            },
            "description": "Provider infrastructure definitions (reusable)",
            "title": "Providers",
            "type": "object"
          },
          "version": {
            "default": "1.0",
            "description": "Configuration schema version",
            "title": "Version",
            "type": "string"
          }
        },
        "title": "LLMConfig",
        "type": "object"
      },
...
      "ProfileConfig": {
        "description": "Profile tier definition (like AWS EC2 instance sizes).\n\nDefines a named configuration that references a provider and specifies\nmodel parameters. Profiles enable right-sizing models for different tasks.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable description of profile use case",
            "title": "Description"
          },
          "max_tokens": {
            "anyOf": [
              {
                "maximum": 128000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum tokens to generate",
            "title": "Max Tokens"
          },
          "model": {
            "description": "Model name (e.g., gpt-4o, claude-sonnet-4, gemini-2.0-flash-exp)",
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "description": "Provider name (must reference a key in providers section)",
            "title": "Provider",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sampling temperature",
            "title": "Temperature"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "title": "ProfileConfig",
        "type": "object"
      },
...
      "ProviderConfig": {
        "description": "Provider infrastructure definition (reusable).\n\nDefines connection details and defaults for an LLM provider endpoint.",
        "properties": {
          "api_key_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Secret key name for API authentication. Maps to WORKFLOW_SECRET_{KEY_NAME} environment variable.",
            "title": "Api Key Secret"
          },
          "api_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "API endpoint URL (optional, uses provider defaults if not specified)",
            "title": "Api Url"
          },
          "api_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "[Azure OpenAI] API version",
            "title": "Api Version"
          },
          "deployment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "[Azure OpenAI] Deployment name",
            "title": "Deployment Name"
          },
          "extra_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom HTTP headers to include in all requests to this provider. Values can use {ENV_VAR} syntax for runtime environment variable resolution.",
            "title": "Extra Headers",
            "type": "object"
          },
          "max_retries": {
            "default": 3,
            "description": "Maximum retry attempts for failed requests",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Retries",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default model for this provider (optional, can be overridden in profile)",
            "title": "Model"
          },
          "retry_delay": {
            "default": 2.0,
            "description": "Initial retry delay in seconds (exponential backoff)",
            "maximum": 60.0,
            "minimum": 0.1,
            "title": "Retry Delay",
            "type": "number"
          },
          "timeout": {
            "default": 60,
            "description": "Request timeout in seconds",
            "maximum": 1800.0,
            "minimum": 1.0,
            "title": "Timeout",
            "type": "integer"
          },
          "type": {
            "description": "Provider type (openai, anthropic, gemini, ollama, azure-openai)",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ProviderConfig",
        "type": "object"
      },
      "RawYAMLRequest": {
        "examples": [
          {
            "raw_yaml": "version: \"1.0\"\nproviders: {}\nprofiles: {}"
          }
        ],
        "properties": {
          "raw_yaml": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "Raw Yaml",
            "type": "string"
          }
        },
        "required": [
          "raw_yaml"
        ],
        "title": "RawYAMLRequest",
        "type": "object"
      },
      "RawYAMLResponse": {
        "properties": {
          "raw_yaml": {
            "title": "Raw Yaml",
            "type": "string"
          }
        },
        "required": [
          "raw_yaml"
        ],
        "title": "RawYAMLResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "in": "cookie",
        "name": "workflows_admin_session",
        "type": "apiKey"
      },
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "workflows-mcp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/v1/filesystem/entries": {
      "get": {
        "operationId": "list_entries_api_admin_v1_filesystem_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Path"
            }
          },
          {
            "in": "query",
            "name": "selection_type",
            "required": false,
            "schema": {
              "default": "folder",
              "enum": [
                "folder",
                "file"
              ],
              "title": "Selection Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "extensions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Extensions"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilesystemEntriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "List Entries",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/llm/config": {
      "get": {
        "operationId": "get_llm_config_api_admin_v1_llm_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Get Llm Config",
        "tags": [
          "admin-v1"
        ]
      },
      "put": {
        "operationId": "put_llm_config_api_admin_v1_llm_config_put",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Put Llm Config",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/llm/export": {
      "get": {
        "operationId": "export_llm_config_api_admin_v1_llm_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawYAMLResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Export Llm Config",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/llm/import": {
      "post": {
        "operationId": "import_llm_config_api_admin_v1_llm_import_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawYAMLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Import Llm Config",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/llm/preview": {
      "post": {
        "operationId": "preview_llm_config_api_admin_v1_llm_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawYAMLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Preview Llm Config",
        "tags": [
          "admin-v1"
        ]
      }
    },
...
    "/api/events/v1/system": {
      "get": {
        "description": "Return minimal SSE stream for v1 events namespace.",
        "operationId": "system_events_api_events_v1_system_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "System Events",
        "tags": [
          "events-v1"
        ]
      }
    },
...
  }
}